*: get MacOS build working
I don't have a MacOS machine (nor can we currently run the tests on
MacOS) but we can at least do a sanity-check by trying to build the
project on MacOS.

Most of the code changes move Linux-only functions (and tests) to
foo_linux{,_test}.go files, while code that was already usable on "unix"
platforms is moved to foo_unix{,_test}.go files. There was also a mild
refactor of some repeated test code, with it all being moved into
testutils. With these minor changes, umoci now builds.
