+++ oci/config/convert/utils_unix.go
@@ -1,0 +1,27 @@
+// +build !linux
+
+/*
+ * umoci: Umoci Modifies Open Containers' Images
+ * Copyright (C) 2016-2019 SUSE LLC.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package convert
+
+// getUnprivilegedMountFlags should return the set of mount flags which are
+// locked by CL_UNPRIVILEGED. However, user namespaces don't exist outside
+// Linux.
+func getUnprivilegedMountFlags(path string) ([]string, error) {
+	return nil, nil
+}
