+++ pkg/testutils/mount_linux.go
@@ -1,3 +1,5 @@
+// +build linux
+
 /*
  * umoci: Umoci Modifies Open Containers' Images
  * Copyright (C) 2016-2020 SUSE LLC
@@ -15,21 +17,21 @@
  * limitations under the License.
  */
 
-package casext
+package testutils
 
 import (
 	"os"
-	"syscall"
 	"testing"
+
+	"golang.org/x/sys/unix"
 )
 
-// readonly makes the given path read-only (by bind-mounting it as "ro").
+// MakeReadOnly makes the given path read-only (by bind-mounting it as "ro").
 // TODO: This should be done through an interface restriction in the test
 //       (which is then backed up by the readonly mount if necessary). The fact
 //       this test is necessary is a sign that we need a better split up of the
 //       CAS interface.
-// Copied from oci/cas/drivers/dir/dir_test.go.
-func readonly(t *testing.T, path string) {
+func MakeReadOnly(t *testing.T, path string) {
 	if os.Geteuid() != 0 {
 		t.Log("readonly tests only work with root privileges")
 		t.Skip()
@@ -37,23 +39,22 @@ func readonly(t *testing.T, path string) {
 
 	t.Logf("mounting %s as readonly", path)
 
-	if err := syscall.Mount(path, path, "", syscall.MS_BIND|syscall.MS_RDONLY, ""); err != nil {
+	if err := unix.Mount(path, path, "", unix.MS_BIND|unix.MS_RDONLY, ""); err != nil {
 		t.Fatalf("mount %s as ro: %s", path, err)
 	}
-	if err := syscall.Mount("none", path, "", syscall.MS_BIND|syscall.MS_REMOUNT|syscall.MS_RDONLY, ""); err != nil {
+	if err := unix.Mount("none", path, "", unix.MS_BIND|unix.MS_REMOUNT|unix.MS_RDONLY, ""); err != nil {
 		t.Fatalf("mount %s as ro: %s", path, err)
 	}
 }
 
-// readwrite undoes the effect of readonly.
-// Copied from oci/cas/drivers/dir/dir_test.go.
-func readwrite(t *testing.T, path string) {
+// MakeReadWrite undos the effect of MakeReadOnly.
+func MakeReadWrite(t *testing.T, path string) {
 	if os.Geteuid() != 0 {
 		t.Log("readonly tests only work with root privileges")
 		t.Skip()
 	}
 
-	if err := syscall.Unmount(path, syscall.MNT_DETACH); err != nil {
+	if err := unix.Unmount(path, unix.MNT_DETACH); err != nil {
 		t.Fatalf("unmount %s: %s", path, err)
 	}
 }
