+++ test/helpers.bash
@@ -36,14 +36,14 @@ mkdir -p "$TESTDIR_TMPDIR"
 # Stores the set of tmpdirs that still have to be cleaned up. Calling
 # teardown_tmpdirs will set this to an empty array (and all the tmpdirs
 # contained within are removed).
-export TESTDIR_LIST="$(mktemp --tmpdir="$TESTDIR_TMPDIR" umoci-integration-tmpdirs.XXXXXX)"
+export TESTDIR_LIST="$(mktemp "$TESTDIR_TMPDIR/umoci-integration-tmpdirs.XXXXXX")"
 
 # setup_tmpdir creates a new temporary directory and returns its name.  Note
 # that if "$ROOTLESS" is true, then removing this tmpdir might be harder than
 # expected -- so tests should not really attempt to clean up tmpdirs.
 function setup_tmpdir() {
 	[[ -n "$UMOCI_TMPDIR" ]] || UMOCI_TMPDIR="$TESTDIR_TMPDIR"
-	mktemp -d --tmpdir="$UMOCI_TMPDIR" umoci-integration-tmpdir.XXXXXXXX | tee -a "$TESTDIR_LIST"
+	mktemp -d "$UMOCI_TMPDIR/umoci-integration-tmpdir.XXXXXXXX" | tee -a "$TESTDIR_LIST"
 }
 
 # setup_tmpdirs just sets up the "built-in" tmpdirs.
